Rename from EunAI to EunAi to follow coding standard
